refactor: create 'signer_registration' module in aggregator
