+++ mithril-aggregator/src/signer_registration/api.rs
@@ -1,0 +1,91 @@
+use async_trait::async_trait;
+
+use mithril_common::{
+    entities::{Epoch, Signer, SignerWithStake, StakeDistribution},
+    StdResult,
+};
+
+use super::SignerRegistrationError;
+
+/// Represents the information needed to handle a signer registration round
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct SignerRegistrationRound {
+    /// Registration round epoch
+    pub epoch: Epoch,
+
+    /// Stake distribution
+    pub(super) stake_distribution: StakeDistribution,
+}
+
+#[cfg(test)]
+impl SignerRegistrationRound {
+    pub(crate) fn dummy(epoch: Epoch, stake_distribution: StakeDistribution) -> Self {
+        Self {
+            epoch,
+            stake_distribution,
+        }
+    }
+}
+
+/// Trait to register a signer
+#[cfg_attr(test, mockall::automock)]
+#[async_trait]
+pub trait SignerRegisterer: Sync + Send {
+    /// Register a signer
+    async fn register_signer(
+        &self,
+        epoch: Epoch,
+        signer: &Signer,
+    ) -> Result<SignerWithStake, SignerRegistrationError>;
+
+    /// Get current open round if exists
+    async fn get_current_round(&self) -> Option<SignerRegistrationRound>;
+}
+
+/// Trait to synchronize signers
+#[cfg_attr(test, mockall::automock)]
+#[async_trait]
+pub trait SignerSynchronizer: Sync + Send {
+    /// Synchronize signers
+    async fn synchronize_signers(
+        &self,
+        epoch: Epoch,
+        signers: &[Signer],
+        stake_distribution: &StakeDistribution,
+    ) -> Result<(), SignerRegistrationError>;
+}
+
+/// Trait to open a signer registration round
+#[cfg_attr(test, mockall::automock)]
+#[async_trait]
+pub trait SignerRegistrationRoundOpener: Sync + Send {
+    /// Open a signer registration round
+    async fn open_registration_round(
+        &self,
+        registration_epoch: Epoch,
+        stake_distribution: StakeDistribution,
+    ) -> StdResult<()>;
+
+    /// Close a signer registration round
+    async fn close_registration_round(&self) -> StdResult<()>;
+}
+
+/// Signer recorder trait
+#[cfg_attr(test, mockall::automock)]
+#[async_trait]
+pub trait SignerRecorder: Sync + Send {
+    /// Record a signer registration
+    async fn record_signer_registration(&self, signer_id: String) -> StdResult<()>;
+}
+
+/// A trait for verifying a [Signer] registration.
+#[cfg_attr(test, mockall::automock)]
+#[async_trait]
+pub trait SignerRegistrationVerifier: Send + Sync {
+    /// Verifies a [Signer] registration.
+    async fn verify(
+        &self,
+        signer: &Signer,
+        stake_distribution: &StakeDistribution,
+    ) -> StdResult<SignerWithStake>;
+}
