+++ mithril-aggregator/src/signer_registration/mod.rs
@@ -1,0 +1,17 @@
+mod api;
+mod errors;
+mod master;
+mod slave;
+mod verifier;
+
+pub use api::{
+    SignerRecorder, SignerRegisterer, SignerRegistrationRound, SignerRegistrationRoundOpener,
+    SignerRegistrationVerifier, SignerSynchronizer,
+};
+pub use errors::SignerRegistrationError;
+pub use master::MithrilSignerRegistererMaster;
+pub use slave::MithrilSignerRegistererSlave;
+pub use verifier::MithrilSignerRegistrationVerifier;
+
+#[cfg(test)]
+pub use api::{MockSignerRecorder, MockSignerRegisterer, MockSignerRegistrationVerifier};
